+++ src/blockchain/esplora/ureq.rs
@@ -112,10 +112,14 @@ impl GetTx for EsploraBlockchain {
     }
 }
 
-impl GetBlockHash for EsploraBlockchain {
+impl GetBlockInfo for EsploraBlockchain {
     fn get_block_hash(&self, height: u64) -> Result<BlockHash, Error> {
         let block_header = self.url_client._get_header(height as u32)?;
         Ok(block_header.block_hash())
+    }
+    fn get_block_header(&self, height: u64) -> Result<BlockHeader, Error> {
+        let block_header = await_or_block!(self.url_client._get_header(height as u32))?;
+        Ok(block_header)
     }
 }